Add support for samples & snippets in pipeline builder
