+++ frontend/packages/console-shared/src/components/editor/yaml-editor-utils.ts
@@ -1,4 +1,6 @@
 import * as URL from 'url';
+import { safeLoad } from 'js-yaml';
+import { saveAs } from 'file-saver';
 import {
   MonacoToProtocolConverter,
   ProtocolToMonacoConverter,
@@ -226,4 +228,18 @@ export const registerYAMLinMonaco = (monaco) => {
   registerYAMLCompletion(LANGUAGE_ID, monaco, m2p, p2m, createDocument, yamlService);
   registerYAMLDocumentSymbols(LANGUAGE_ID, monaco, p2m, createDocument, yamlService);
   registerYAMLHover(LANGUAGE_ID, monaco, m2p, p2m, createDocument, yamlService);
+};
+
+export const downloadYaml = (data) => {
+  const blob = new Blob([data], { type: 'text/yaml;charset=utf-8' });
+  let filename = 'k8s-object.yaml';
+  try {
+    const obj = safeLoad(data);
+    if (obj.kind) {
+      filename = `${obj.kind.toLowerCase()}-${obj.metadata.name}.yaml`;
+    }
+  } catch (unused) {
+    // unused
+  }
+  saveAs(blob, filename);
 };
