+++ frontend/packages/console-shared/src/utils/index.ts
@@ -15,3 +15,4 @@ export * from './alert-utils';
 export * from './operator-utils';
 export * from './helm-utils';
 export * from './hpa-utils';
+export * from './sample-utils';
